requirement.txt
matplotlib-inline==0.1.6
missingno==0.5.2
numpy==1.23.5
pandas==1.5.3
pytorch-lightning==1.5.10
scikit-learn==1.1.3
scipy==1.9.3
seaborn==0.12.2
torch==1.12.1
torch-cluster==1.6.0
torch-geometric==2.1.0
torch-scatter==2.1.0
torch-sparse==0.6.15
torch-spatiotemporal==0.1.1
torchaudio==0.12.1
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.13.1

